Add esdoc.url to package.json
Field contains a lodash string template accepting `version` and
rendering the base url of the published ESDoc output for the package at
that version.

Other packages that depend on this one can then grab that string
template, render in the version they're looking for, and use that base
url to generate ESDoc @external definitions for identifiers in this
package, avoiding having to keep the definitions up-to-date by hand.

This is the first implementation of this process, so it is subject to
change, but it seems alright for a start.