Add toAbsPath, toAbsDir, toResDir functions to path
And some tests... could use more, and should have toResPath for
completeness. This came from needs in @nrser/ugh - i have a path, and i
need to be able to assemble an absolute path from it WITHOUT resolving
off the working dir (since we're all over the place directory-wise in
there with includes and stuff).